requirements.txt
pypdf
openai
faiss-cpu
python-dotenv
fastapi
uvicorn
sentence-transformers
PyPDF2

